AM-929: implementation of replace existing true to handle duplicate records.
